Added test for design temp calculator tool to rules/test_rules_engine folder, cleaned up design_temp_calc.py file. Also verfied and test case passes for this file.
