--- design_temp/test_design_temp_calc.py
@@ -1,4 +1,0 @@
-from datetime import date
-from design_temp_calc import get_date_range
-
-# make unit tests for parts of the design_temp_calc
